Cleanup old code.
