+++ peering/fields.py
@@ -1,9 +1,5 @@
-from ipaddress import ip_network
-
 from django.core.validators import MinValueValidator, MaxValueValidator, RegexValidator
 from django.db import models
-
-from .formfields import IPNetworkFormField
 
 
 class ASNField(models.BigIntegerField):
